backend/apps/product/serializers.py
from django.db.models import fields
from rest_framework import serializers
from .models import Category, Product


class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = (
            "id",
            "nom",
            "get_absolute_url",
            "description",
            "prix",
            "get_image",
            "get_thumbnail"
        )


class CategorySerializer(serializers.ModelSerializer):
    products = ProductSerializer(many=True)

    class Meta:
        model = Category
        fields = (
            "id",
            "nom",
            "get_absolute_url",
            "products"
        )
