backend/requirements.txt
autopep8==1.5.7               
attrs==20.3.0
backcall==0.2.0
chardet==4.0.0
decorator==4.0.6
django-environ==0.4.5
django-extensions==3.1.3
django-redis==4.12.1
flake8==3.9.1
flake8-commas==2.0.0
flake8-mypy==17.8.0
gevent==21.1.2
greenlet==1.0.0
gunicorn==20.1.0
ipython==7.27.0
ipython-genutils==0.2.0
jedi==0.18.0
mccabe==0.6.1
mypy==0.812
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.8.1
pytz==2021.1
redis==3.5.3
rope==0.19.0
simplegeneric==0.8.1
traitlets==5.0.5
typed-ast==1.4.3
wcwidth==0.2.5
Werkzeug==1.0.1
sentry-sdk==1.1.0
djangorestframework==3.12.4





asgiref==3.4.1                              
astroid==2.7.3
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.8.0
django-heroku==0.3.1
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==3.2
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
oauthlib==3.1.1
Pillow==8.3.2
platformdirs==2.3.0
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
pylint==2.10.2
python3-openid==3.2.0
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
stripe==2.60.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.3.0
wrapt==1.12.1
